project completion

+++ todolist_app/views.py
@@ -27,7 +27,10 @@ def todolist(request):
 @login_required
 def delete_task(request, task_id):
         task = TaskList.objects.get(pk=task_id)
-        task.delete()
+        if task.manage == request.user:
+            task.delete()
+        else:
+            messages.error(request,("Access Restricted, You Are Not Allowed."))            
         return redirect('todolist')
    
 # Edit tasks 
@@ -48,8 +51,12 @@ def edit_task(request, task_id):
 @login_required    
 def complete_task(request, task_id):
         task = TaskList.objects.get(pk=task_id)
-        task.done = True
-        task.save()
+        if task.manage == request.user:
+            task.done = True
+            task.save()
+        else:
+            messages.error(request,("Access Restricted, You Are Not Allowed."))
+                
         return redirect(request.META.get('HTTP_REFERER', 'todolist'))
 
 @login_required
